Fix cast when generating sign mask

+++ src/fpx.c
@@ -25,7 +25,7 @@ int8_t ct_compare(const uint8_t *a, const uint8_t *b, unsigned int len)
     for (unsigned int i = 0; i < len; i++)
         r |= a[i] ^ b[i];
 
-    return (-(int8_t)r) >> (8*sizeof(uint8_t)-1);
+    return (int8_t)((-(int32_t)r) >> (8*sizeof(uint32_t)-1));
 }
 
 
@@ -1046,7 +1046,7 @@ int8_t cmp_f2elm(const f2elm_t x, const f2elm_t y)
     for (int i = NWORDS_FIELD-1; i >= 0; i--)
         r |= (a[0][i] ^ b[0][i]) | (a[1][i] ^ b[1][i]);
 
-    return (-(int8_t)r) >> (8*sizeof(uint8_t)-1);
+    return (int8_t)((-(int32_t)r) >> (8*sizeof(uint32_t)-1));
 }
